Move to ppc/ folder
Stick with 3 letter architectures nameing scheme.
