+++ include/boost/simd/arch/ppc/spec.hpp
@@ -2,18 +2,18 @@
 /*!
   @file
 
-  Aggregate SIMD numerical and type limits for PPC
+  Gather informations for PowerPC SIMD extensions
 
-  @copyright 2012 - 2015 NumScale SAS
+  @copyright 2016 NumScale SAS
 
   Distributed under the Boost Software License, Version 1.0.
   (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
 
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_ARCH_POWER_LIMITS_HPP_INCLUDED
-#define BOOST_SIMD_ARCH_POWER_LIMITS_HPP_INCLUDED
+#ifndef BOOST_SIMD_ARCH_PPC_SPEC_HPP_INCLUDED
+#define BOOST_SIMD_ARCH_PPC_SPEC_HPP_INCLUDED
 
-#include <boost/simd/arch/power/vmx/limits.hpp>
+#include <boost/simd/arch/ppc/vmx/spec.hpp>
 
 #endif
