+++ include/boost/simd/arch/ppc/vmx/spec.hpp
@@ -11,8 +11,8 @@
 
 **/
 //==================================================================================================
-#ifndef BOOST_SIMD_ARCH_POWER_VMX_SPEC_HPP_INCLUDED
-#define BOOST_SIMD_ARCH_POWER_VMX_SPEC_HPP_INCLUDED
+#ifndef BOOST_SIMD_ARCH_PPC_VMX_SPEC_HPP_INCLUDED
+#define BOOST_SIMD_ARCH_PPC_VMX_SPEC_HPP_INCLUDED
 
 #include <boost/simd/detail/predef.hpp>
 
@@ -52,7 +52,7 @@
 #define __bool bool
 #endif
 
-#include <boost/simd/arch/power/vmx/as_simd.hpp>
+#include <boost/simd/arch/ppc/vmx/as_simd.hpp>
 
 #endif
 #endif
